+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 Python PEP8 Autoformat
 
+## 2013.05.09-01
+  - FIX error parsing "ignore" and "select" settings.
+  - Changes misinterpreted "aggressive" settings from boolean to counter.
+  - ADD new boolean settings "list-fixes".
+
 ## 2013.03.25-02
   - FIX [Issue #7 Does not work with syntax "Python Django"](https://bitbucket.org/StephaneBunel/pythonpep8autoformat/issue/10/does-not-work-with-syntax-python-django)